requirements.txt
mysql-connector-python
pandas
numpy
matplotlib
seaborn
